Add command system functionalities and remove obsolete commands
- Implemented AliasCommand for creating, listing, and removing command aliases.
- Added ClearCommand to clear the terminal screen.
- Introduced HelpCommand to display help information for available commands.
- Created PSCommand to list running processes on the system.
- Developed QuitCommand for exiting the game with optional saving.
- Removed TouchCommand and WcCommand as they were deemed unnecessary.
- Enhanced ExploitCommand and VulnScanCommand for better vulnerability management.
- Added VulnsCommand to display the user's vulnerability database.
