+++ Assets/Scripts/Core/CommandSystem/Commands/CLI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 878f2d050a86000b1857544630d66d5d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
