+++ Assets/Scripts/Core/CommandSystem/Commands/FileOps.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1529baf0655cab9a9a9502ba24801dd9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
