+++ Assets/Scripts/Core/CommandSystem/Commands/Networking/SshCommand.cs
@@ -1,58 +1,63 @@
 using System;
+using SampleOS.Core.CommandSystem;
 
-public class SshCommand : ICommand
+namespace SampleOS.Core.CommandSystem.Commands.Networking
 {
-  private VirtualNetwork network;
-  private CommandProcessor commandProcessor;
 
-  public string Name => "ssh";
-  public string Description => "Connect to a remote system";
-  public string Usage => "ssh [username@]hostname";
-
-  public SshCommand(VirtualNetwork network, CommandProcessor commandProcessor)
+  public class SshCommand : ICommand
   {
-    this.network = network;
-    this.commandProcessor = commandProcessor;
-  }
+    private VirtualNetwork network;
+    private CommandProcessor commandProcessor;
 
-  public void Execute(string[] args, ITerminalOutput output)
-  {
-    if (args.Length == 0)
+    public string Name => "ssh";
+    public string Description => "Connect to a remote system";
+    public string Usage => "ssh [username@]hostname";
+
+    public SshCommand(VirtualNetwork network, CommandProcessor commandProcessor)
     {
-      output.AppendText($"Usage: {Usage}\n");
-      return;
+      this.network = network;
+      this.commandProcessor = commandProcessor;
     }
 
-    string target = args[0];
-    string username = "user";
-    string host = target;
-
-    // Parse username@host format
-    if (target.Contains("@"))
+    public void Execute(string[] args, ITerminalOutput output)
     {
-      string[] parts = target.Split('@');
-      username = parts[0];
-      host = parts[1];
-    }
+      if (args.Length == 0)
+      {
+        output.AppendText($"Usage: {Usage}\n");
+        return;
+      }
 
-    // For now, no password handling
-    string password = "";
+      string target = args[0];
+      string username = "user";
+      string host = target;
 
-    output.AppendText($"Connecting to {host} as {username}...\n");
+      // Parse username@host format
+      if (target.Contains("@"))
+      {
+        string[] parts = target.Split('@');
+        username = parts[0];
+        host = parts[1];
+      }
 
-    var result = network.Connect(host, username, password);
+      // For now, no password handling
+      string password = "";
 
-    if (result.IsSuccess)
-    {
-      RemoteSystem system = result.Data;
-      output.AppendText($"Connected to {system.Name}\n");
+      output.AppendText($"Connecting to {host} as {username}...\n");
 
-      // Switch to the remote system's filesystem
-      commandProcessor.SetFileSystem(system.FileSystem);
-    }
-    else
-    {
-      output.AppendText($"Connection failed: {result.ErrorMessage}\n");
+      var result = network.Connect(host, username, password);
+
+      if (result.IsSuccess)
+      {
+        RemoteSystem system = result.Data;
+        output.AppendText($"Connected to {system.Name}\n");
+
+        // Switch to the remote system's filesystem
+        commandProcessor.SetFileSystem(system.FileSystem);
+      }
+      else
+      {
+        output.AppendText($"Connection failed: {result.ErrorMessage}\n");
+      }
     }
   }
 }
