+++ Assets/Scripts/Core/CommandSystem/Commands/Systems.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3b83bbec744c451b59826181d40a09d0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
