+++ Assets/Scripts/Core/CommandSystem/Commands/Systems/ClearCommand.cs
@@ -1,0 +1,21 @@
+using SampleOS.Core.CommandSystem;
+
+namespace SampleOS.Core.CommandSystem.Commands.Systems
+{
+
+  public class ClearCommand : ICommand
+  {
+    public string Name => "clear";
+    public string Description => "Clear the terminal screen";
+    public string Usage => "clear";
+
+    public ClearCommand()
+    {
+    }
+
+    public void Execute(string[] args, ITerminalOutput output)
+    {
+      output.Clear();
+    }
+  }
+}
